requirements.txt
python-dotenv
openai==0.28
